fix(perps): resolve circular dependency in utils re-export
Move buildPerpsVipTrackingData to its own file (utils/trackingData.ts)
and re-export directly from utils.ts, bypassing the barrel that creates
a cycle through transactionTransforms.ts.
